Dynapse: cleanup, fix warnings.

+++ include/devices/dynapse.h
@@ -639,7 +639,7 @@ bool caerDynapseWriteSram(caerDeviceHandle handle, uint16_t coreId, uint32_t neu
 *
 * @return true on success, false otherwise
 */
-bool caerDynapseSendDataToUSB(caerDeviceHandle handle, const uint32_t *data, int numConfig);
+bool caerDynapseSendDataToUSB(caerDeviceHandle handle, const uint32_t *data, size_t numConfig);
 
 /*
  * Remember to Select the chip before calling this function
